fix connect side
