+++ dev/sparktestsupport/modules.py
@@ -716,12 +716,12 @@ pyspark_pandas_slow = Module(
         "pyspark.pandas.tests.indexes.test_base",
         "pyspark.pandas.tests.indexes.test_datetime",
         # "splits of pyspark.pandas.tests.test_dataframe",
-        "pyspark.pandas.tests.test_dataframe_splits.split_a_b",
-        "pyspark.pandas.tests.test_dataframe_splits.split_c",
-        "pyspark.pandas.tests.test_dataframe_splits.split_d_k",
-        "pyspark.pandas.tests.test_dataframe_splits.split_l_p",
-        "pyspark.pandas.tests.test_dataframe_splits.split_q_r",
-        "pyspark.pandas.tests.test_dataframe_splits.split_s_x",
+        "pyspark.pandas.tests.test_dataframe_splits.test_dataframe_split_a_b",
+        "pyspark.pandas.tests.test_dataframe_splits.test_dataframe_split_c",
+        "pyspark.pandas.tests.test_dataframe_splits.test_dataframe_split_d_k",
+        "pyspark.pandas.tests.test_dataframe_splits.test_dataframe_split_l_p",
+        "pyspark.pandas.tests.test_dataframe_splits.test_dataframe_split_q_r",
+        "pyspark.pandas.tests.test_dataframe_splits.test_dataframe_split_s_z",
         "pyspark.pandas.tests.test_groupby",
         "pyspark.pandas.tests.test_groupby_slow",
         "pyspark.pandas.tests.test_indexing",
@@ -877,8 +877,13 @@ pyspark_pandas_slow_connect = Module(
     python_test_goals=[
         # pandas-on-Spark unittests
         "pyspark.pandas.tests.connect.indexes.test_parity_datetime",
-        "pyspark.pandas.tests.connect.test_parity_dataframe",
-        "pyspark.pandas.tests.connect.test_parity_dataframe_slow",
+        # "splits of pyspark.pandas.tests.connect.test_parity_dataframee",
+        "pyspark.pandas.tests.connect.test_parity_dataframe_splits.test_parity_dataframe_split_a_b",
+        "pyspark.pandas.tests.connect.test_parity_dataframe_splits.test_parity_dataframe_split_c",
+        "pyspark.pandas.tests.connect.test_parity_dataframe_splits.test_parity_dataframe_split_d_k",
+        "pyspark.pandas.tests.connect.test_parity_dataframe_splits.test_parity_dataframe_split_l_p",
+        "pyspark.pandas.tests.connect.test_parity_dataframe_splits.test_parity_dataframe_split_q_r",
+        "pyspark.pandas.tests.connect.test_parity_dataframe_splits.test_parity_dataframe_split_s_z",
         "pyspark.pandas.tests.connect.test_parity_groupby",
         "pyspark.pandas.tests.connect.test_parity_groupby_slow",
         "pyspark.pandas.tests.connect.test_parity_indexing",
