+++ python/pyspark/pandas/tests/connect/test_parity_dataframe.py
@@ -22,6 +22,9 @@ from pyspark.testing.connectutils import ReusedConnectTestCase
 from pyspark.testing.pandasutils import PandasOnSparkTestUtils
 
 
+# We should not add this test case to CI ('dev/sparktestsupport/modules.py'),
+# instead we divide it into several splits in 'test_parity_dataframe_splits' and so that the testing
+# framework could be able to execute them in parallel.
 class DataFrameParityTests(DataFrameTestsMixin, PandasOnSparkTestUtils, ReusedConnectTestCase):
     @property
     def psdf(self):
@@ -140,6 +143,90 @@ class DataFrameParityTests(DataFrameTestsMixin, PandasOnSparkTestUtils, ReusedCo
     )
     def test_append(self):
         super().test_append()
+
+    @unittest.skip(
+        "TODO(SPARK-43610): Enable `InternalFrame.attach_distributed_column` in Spark Connect."
+    )
+    def test_at_time(self):
+        super().test_at_time()
+
+    @unittest.skip(
+        "TODO(SPARK-43611): Fix unexpected `AnalysisException` from Spark Connect client."
+    )
+    def test_backfill(self):
+        super().test_backfill()
+
+    @unittest.skip(
+        "TODO(SPARK-43610): Enable `InternalFrame.attach_distributed_column` in Spark Connect."
+    )
+    def test_between_time(self):
+        super().test_between_time()
+
+    @unittest.skip(
+        "TODO(SPARK-43611): Fix unexpected `AnalysisException` from Spark Connect client."
+    )
+    def test_bfill(self):
+        super().test_bfill()
+
+    @unittest.skip(
+        "TODO(SPARK-43613): Enable pyspark.pandas.spark.functions.covar in Spark Connect."
+    )
+    def test_cov(self):
+        super().test_cov()
+
+    @unittest.skip(
+        "TODO(SPARK-43611): Fix unexpected `AnalysisException` from Spark Connect client."
+    )
+    def test_diff(self):
+        super().test_diff()
+
+    @unittest.skip("TODO(SPARK-43615): Enable DataFrameSlowParityTests.test_eval.")
+    def test_eval(self):
+        super().test_eval()
+
+    @unittest.skip(
+        "TODO(SPARK-43611): Fix unexpected `AnalysisException` from Spark Connect client."
+    )
+    def test_ffill(self):
+        super().test_ffill()
+
+    @unittest.skip(
+        "TODO(SPARK-43616): Enable pyspark.pandas.spark.functions.mode in Spark Connect."
+    )
+    def test_mode(self):
+        super().test_mode()
+
+    @unittest.skip(
+        "TODO(SPARK-43611): Fix unexpected `AnalysisException` from Spark Connect client."
+    )
+    def test_pad(self):
+        super().test_pad()
+
+    @unittest.skip(
+        "TODO(SPARK-43611): Fix unexpected `AnalysisException` from Spark Connect client."
+    )
+    def test_pct_change(self):
+        super().test_pct_change()
+
+    @unittest.skip(
+        "TODO(SPARK-43617): Enable pyspark.pandas.spark.functions.product in Spark Connect."
+    )
+    def test_product(self):
+        super().test_product()
+
+    @unittest.skip("TODO(SPARK-43618): Fix pyspark.sq.column._unary_op to work with Spark Connect.")
+    def test_rank(self):
+        super().test_rank()
+
+    @unittest.skip(
+        "TODO(SPARK-43611): Fix unexpected `AnalysisException` from Spark Connect client."
+    )
+    def test_shift(self):
+        super().test_shift()
+
+    @unittest.skip("TODO(SPARK-43619): Enable DataFrameSlowParityTests.test_udt.")
+    def test_udt(self):
+        super().test_udt()
 
 
 if __name__ == "__main__":
